refactor tests
refactor test cases
