+++ src/test/java/org/cicirello/search/evo/BasePopulationTests.java
@@ -22,18 +22,12 @@ package org.cicirello.search.evo;
 
 import org.junit.jupiter.api.*;
 import static org.junit.jupiter.api.Assertions.*;
-import org.cicirello.util.Copyable;
-import org.cicirello.search.operators.Initializer;
-import org.cicirello.search.SolutionCostPair;
 import org.cicirello.search.ProgressTracker;
-import org.cicirello.search.problems.Problem;
-import org.cicirello.search.problems.OptimizationProblem;
-import org.cicirello.search.problems.IntegerCostOptimizationProblem;
 
 /**
  * JUnit test cases for BasePopulation.
  */
-public class BasePopulationTests {
+public class BasePopulationTests extends SharedTestPopulations {
 	
 	private static final double EPSILON = 1e-10;
 	
@@ -116,91 +110,7 @@ public class BasePopulationTests {
 			tracker, 
 			3
 		);
-		assertTrue(tracker == pop.getProgressTracker());
-		tracker = new ProgressTracker<TestObject>();
-		pop.setProgressTracker(tracker);
-		assertTrue(tracker == pop.getProgressTracker());
-		
-		pop.init();
-		assertEquals(10, pop.size());
-		assertEquals(7, pop.mutableSize());
-		assertEquals(6.4, pop.getFitnessOfMostFit(), EPSILON);
-		assertEquals(1.0/7.0, pop.getMostFit().getCostDouble(), EPSILON);
-		assertEquals(6, pop.getMostFit().getSolution().id);
-		assertEquals(tracker.getSolution(), pop.getMostFit().getSolution());
-		int[] expected = { 2, 3, 4, 5, 6, 5, 4, 3, 2, 1};
-		for (int i = 0; i < 10; i++) {
-			// fitnesses of original before selection.
-			assertEquals(expected[9-i]+0.4, pop.getFitness(i), EPSILON);
-		}
-		assertFalse(selection.called);
-		pop.select();
-		assertTrue(selection.called);
-		for (int i = 0; i < 10; i++) {
-			// fitnesses of original before selection.
-			assertEquals(expected[9-i]+0.4, pop.getFitness(i), EPSILON);
-			if (i < 7) {
-				// subject to mutation to opposite order since we selected, which reversed.
-				assertEquals(expected[i+3], pop.get(i).id);
-			} 
-		}
-		assertEquals(6.4, pop.getFitnessOfMostFit(), EPSILON);
-		assertEquals(1.0/7.0, pop.getMostFit().getCostDouble(), EPSILON);
-		assertEquals(6, pop.getMostFit().getSolution().id);
-		assertEquals(tracker.getSolution(), pop.getMostFit().getSolution());
-		pop.replace();
-		int[] expectedNow = { 5, 6, 5, 4, 3, 2, 1, 4, 6, 5 };
-		for (int i = 0; i < 10; i++) {
-			assertEquals(expectedNow[i]+0.4, pop.getFitness(i), EPSILON, "index i="+i);
-		}
-		pop.select();
-		for (int i = 0; i < 7; i++) {
-			assertEquals(expectedNow[6-i], pop.get(i).id);
-		}
-		assertEquals(6.4, pop.getFitnessOfMostFit(), EPSILON);
-		assertEquals(1.0/7.0, pop.getMostFit().getCostDouble(), EPSILON);
-		assertEquals(6, pop.getMostFit().getSolution().id);
-		assertEquals(tracker.getSolution(), pop.getMostFit().getSolution());
-		
-		f.changeFitness(1);
-		pop.updateFitness(1);
-		f.changeFitness(10);
-		pop.updateFitness(2);
-		pop.replace();
-		assertEquals(2+0.4+1, pop.getFitness(1), EPSILON);
-		assertEquals(3+0.4+10, pop.getFitness(2), EPSILON);
-		assertEquals(13.4, pop.getFitnessOfMostFit(), EPSILON);
-		assertEquals(3, pop.getMostFit().getSolution().id);
-		assertEquals(1.0/4.0, pop.getMostFit().getCostDouble(), EPSILON);
-		
-		int[] andNow = {1, 2, 3, 4, 5, 6, 5, 4, 6, 5};
-		double[] andNowFitness = {1.4, 3.4, 13.4, 4.4, 5.4, 6.4, 5.4, 4.4, 6.4, 5.4};
-		for (int i = 0; i < 10; i++) {
-			assertEquals(andNowFitness[i], pop.getFitness(i), EPSILON, "index i="+i);
-		}
-		
-		f.changeFitness(12);
-		BasePopulation.Double<TestObject> pop2 = pop.split();
-		
-		// original should be same after split
-		for (int i = 0; i < 10; i++) {
-			assertEquals(andNowFitness[i], pop.getFitness(i), EPSILON, "index i="+i);
-		}
-		
-		// trackers should be same
-		assertTrue(pop.getProgressTracker() == pop2.getProgressTracker());
-		
-		pop2.init();
-		assertEquals(10, pop2.size());
-		assertEquals(7, pop2.mutableSize());
-		for (int i = 0; i < 10; i++) {
-			assertEquals(1-i+12+0.4, pop2.getFitness(i), EPSILON);
-		}
-		
-		// Original should be unchanged after split copy does stuff
-		for (int i = 0; i < 10; i++) {
-			assertEquals(andNowFitness[i], pop.getFitness(i), EPSILON, "index i="+i);
-		}
+		verifyDouble(pop, f, tracker, selection, p -> ((BasePopulation.Double<TestObject>)p).getFitnessOfMostFit(), 3);
 	}
 	
 	@Test
@@ -217,90 +127,7 @@ public class BasePopulationTests {
 			tracker, 
 			0
 		);
-		assertTrue(tracker == pop.getProgressTracker());
-		tracker = new ProgressTracker<TestObject>();
-		pop.setProgressTracker(tracker);
-		assertTrue(tracker == pop.getProgressTracker());
-		
-		pop.init();
-		assertEquals(10, pop.size());
-		assertEquals(10, pop.mutableSize());
-		assertEquals(6.4, pop.getFitnessOfMostFit(), EPSILON);
-		assertEquals(1.0/7.0, pop.getMostFit().getCostDouble(), EPSILON);
-		assertEquals(6, pop.getMostFit().getSolution().id);
-		assertEquals(tracker.getSolution(), pop.getMostFit().getSolution());
-		int[] expected = { 2, 3, 4, 5, 6, 5, 4, 3, 2, 1};
-		for (int i = 0; i < 10; i++) {
-			// fitnesses of original before selection.
-			assertEquals(expected[9-i]+0.4, pop.getFitness(i), EPSILON);
-		}
-		assertFalse(selection.called);
-		pop.select();
-		assertTrue(selection.called);
-		for (int i = 0; i < 10; i++) {
-			// fitnesses of original before selection.
-			assertEquals(expected[9-i]+0.4, pop.getFitness(i), EPSILON);
-			// subject to mutation to opposite order since we selected, which reversed.
-			assertEquals(expected[i], pop.get(i).id);
-			
-		}
-		assertEquals(6.4, pop.getFitnessOfMostFit(), EPSILON);
-		assertEquals(1.0/7.0, pop.getMostFit().getCostDouble(), EPSILON);
-		assertEquals(6, pop.getMostFit().getSolution().id);
-		assertEquals(tracker.getSolution(), pop.getMostFit().getSolution());
-		pop.replace();
-		pop.select();
-		for (int i = 0; i < 10; i++) {
-			assertEquals(expected[9-i], pop.get(i).id);
-			assertEquals(expected[i]+0.4, pop.getFitness(i), EPSILON);
-		}
-		assertEquals(6.4, pop.getFitnessOfMostFit(), EPSILON);
-		assertEquals(1.0/7.0, pop.getMostFit().getCostDouble(), EPSILON);
-		assertEquals(6, pop.getMostFit().getSolution().id);
-		assertEquals(tracker.getSolution(), pop.getMostFit().getSolution());
-		
-		f.changeFitness(1);
-		pop.updateFitness(0);
-		f.changeFitness(10);
-		pop.updateFitness(1);
-		pop.replace();
-		assertEquals(expected[9]+0.4+1, pop.getFitness(0), EPSILON);
-		assertEquals(expected[8]+0.4+10, pop.getFitness(1), EPSILON);
-		assertEquals(12.4, pop.getFitnessOfMostFit(), EPSILON);
-		assertEquals(2, pop.getMostFit().getSolution().id);
-		assertEquals(1.0/3.0, pop.getMostFit().getCostDouble(), EPSILON);
-		
-		f.changeFitness(12);
-		BasePopulation.Double<TestObject> pop2 = pop.split();
-		
-		// orginal should be same
-		assertEquals(expected[9]+0.4+1, pop.getFitness(0), EPSILON);
-		assertEquals(expected[8]+0.4+10, pop.getFitness(1), EPSILON);
-		assertEquals(12.4, pop.getFitnessOfMostFit(), EPSILON);
-		assertEquals(2, pop.getMostFit().getSolution().id);
-		assertEquals(1.0/3.0, pop.getMostFit().getCostDouble(), EPSILON);
-		
-		// trackers should be same
-		assertTrue(pop.getProgressTracker() == pop2.getProgressTracker());
-		
-		pop2.init();
-		assertEquals(10, pop2.size());
-		assertEquals(10, pop2.mutableSize());
-		for (int i = 0; i < 10; i++) {
-			assertEquals(1-i+12+0.4, pop2.getFitness(i), EPSILON);
-		}
-		
-		assertFalse(pop.evolutionIsPaused());
-		assertFalse(pop2.evolutionIsPaused());
-		tracker.stop();
-		assertTrue(pop.evolutionIsPaused());
-		assertTrue(pop2.evolutionIsPaused());
-		tracker.start();
-		assertFalse(pop.evolutionIsPaused());
-		assertFalse(pop2.evolutionIsPaused());
-		tracker.update(0.0, new TestObject(), true);
-		assertTrue(pop.evolutionIsPaused());
-		assertTrue(pop2.evolutionIsPaused());
+		verifyDouble(pop, f, tracker, selection, p -> ((BasePopulation.Double<TestObject>)p).getFitnessOfMostFit(), 0);
 	}
 	
 	@Test
@@ -317,22 +144,7 @@ public class BasePopulationTests {
 			tracker, 
 			0
 		);
-		pop.init();
-		pop.select();
-		TestObject[] firstSelect = new TestObject[10];
-		for (int i = 0; i < 10; i++) {
-			firstSelect[i] = pop.get(i);
-		}
-		pop.replace();
-		pop.select();
-		TestObject[] secondSelect = new TestObject[10];
-		for (int i = 0; i < 10; i++) {
-			secondSelect[i] = pop.get(i);
-		}
-		for (int i = 0; i < 10; i++) {
-			assertFalse(firstSelect[i] == secondSelect[9-i]);
-			assertEquals(firstSelect[i], secondSelect[9-i]);
-		}
+		verifySelectCopies(pop);
 	}
 	
 	@Test
@@ -449,22 +261,7 @@ public class BasePopulationTests {
 			tracker, 
 			0
 		);
-		pop.init();
-		pop.select();
-		TestObject[] firstSelect = new TestObject[10];
-		for (int i = 0; i < 10; i++) {
-			firstSelect[i] = pop.get(i);
-		}
-		pop.replace();
-		pop.select();
-		TestObject[] secondSelect = new TestObject[10];
-		for (int i = 0; i < 10; i++) {
-			secondSelect[i] = pop.get(i);
-		}
-		for (int i = 0; i < 10; i++) {
-			assertFalse(firstSelect[i] == secondSelect[9-i]);
-			assertEquals(firstSelect[i], secondSelect[9-i]);
-		}
+		verifySelectCopies(pop);
 	}
 	
 	@Test
@@ -481,86 +278,7 @@ public class BasePopulationTests {
 			tracker, 
 			3
 		);
-		assertTrue(tracker == pop.getProgressTracker());
-		tracker = new ProgressTracker<TestObject>();
-		pop.setProgressTracker(tracker);
-		assertTrue(tracker == pop.getProgressTracker());
-		
-		pop.init();
-		assertEquals(10, pop.size());
-		assertEquals(7, pop.mutableSize());
-		assertEquals(16, pop.getFitnessOfMostFit());
-		assertEquals(94, pop.getMostFit().getCost());
-		assertEquals(6, pop.getMostFit().getSolution().id);
-		assertEquals(tracker.getSolution(), pop.getMostFit().getSolution());
-		int[] expected = { 2, 3, 4, 5, 6, 5, 4, 3, 2, 1};
-		for (int i = 0; i < 10; i++) {
-			// fitnesses of original before selection.
-			assertEquals(expected[9-i]+10, pop.getFitness(i));
-		}
-		assertFalse(selection.called);
-		pop.select();
-		assertTrue(selection.called);
-		for (int i = 0; i < 10; i++) {
-			// fitnesses of original before selection.
-			assertEquals(expected[9-i]+10, pop.getFitness(i));
-			if (i < 7) {
-				// subject to mutation to opposite order since we selected, which reversed.
-				assertEquals(expected[i+3], pop.get(i).id);
-			} 
-		}
-		assertEquals(16, pop.getFitnessOfMostFit());
-		assertEquals(94, pop.getMostFit().getCost());
-		assertEquals(6, pop.getMostFit().getSolution().id);
-		assertEquals(tracker.getSolution(), pop.getMostFit().getSolution());
-		
-		f.changeFitness(10);
-		pop.updateFitness(4);
-		
-		pop.replace();
-		int[] expectedNow = { 5, 6, 5, 4, 3, 2, 1, 4, 6, 5 };
-		for (int i = 0; i < 10; i++) {
-			if (i!=4) assertEquals(expectedNow[i]+10, pop.getFitness(i), "index i="+i);
-			else assertEquals(expectedNow[i]+20, pop.getFitness(i), "index i="+i);
-		}
-		pop.select();
-		for (int i = 0; i < 7; i++) {
-			assertEquals(expectedNow[6-i], pop.get(i).id);
-		}
-		assertEquals(23, pop.getFitnessOfMostFit());
-		assertEquals(97, pop.getMostFit().getCost());
-		assertEquals(3, pop.getMostFit().getSolution().id);
-		
-		f.changeFitness(12);
-		BasePopulation.Integer<TestObject> pop2 = pop.split();
-		
-		// original should be same after split
-		for (int i = 0; i < 10; i++) {
-			if (i!=4) assertEquals(expectedNow[i]+10, pop.getFitness(i), "index i="+i);
-			else assertEquals(expectedNow[i]+20, pop.getFitness(i), "index i="+i);
-		}
-		for (int i = 0; i < 7; i++) {
-			assertEquals(expectedNow[6-i], pop.get(i).id);
-		}
-		
-		// trackers should be same
-		assertTrue(pop.getProgressTracker() == pop2.getProgressTracker());
-		
-		pop2.init();
-		assertEquals(10, pop2.size());
-		assertEquals(7, pop2.mutableSize());
-		for (int i = 0; i < 10; i++) {
-			assertEquals(1-i+12+10, pop2.getFitness(i));
-		}
-		
-		// Original should be unchanged after split copy does stuff
-		for (int i = 0; i < 10; i++) {
-			if (i!=4) assertEquals(expectedNow[i]+10, pop.getFitness(i), "index i="+i);
-			else assertEquals(expectedNow[i]+20, pop.getFitness(i), "index i="+i);
-		}
-		for (int i = 0; i < 7; i++) {
-			assertEquals(expectedNow[6-i], pop.get(i).id);
-		}
+		verifyInteger(pop, f, tracker, selection, p -> ((BasePopulation.Integer<TestObject>)p).getFitnessOfMostFit(), 3);
 	}
 	
 	@Test
@@ -577,90 +295,7 @@ public class BasePopulationTests {
 			tracker, 
 			0
 		);
-		assertTrue(tracker == pop.getProgressTracker());
-		tracker = new ProgressTracker<TestObject>();
-		pop.setProgressTracker(tracker);
-		assertTrue(tracker == pop.getProgressTracker());
-		
-		pop.init();
-		assertEquals(10, pop.size());
-		assertEquals(10, pop.mutableSize());
-		assertEquals(16, pop.getFitnessOfMostFit());
-		assertEquals(94, pop.getMostFit().getCost());
-		assertEquals(6, pop.getMostFit().getSolution().id);
-		assertEquals(tracker.getSolution(), pop.getMostFit().getSolution());
-		int[] expected = { 2, 3, 4, 5, 6, 5, 4, 3, 2, 1};
-		for (int i = 0; i < 10; i++) {
-			// fitnesses of original before selection.
-			assertEquals(expected[9-i]+10, pop.getFitness(i));
-		}
-		assertFalse(selection.called);
-		pop.select();
-		assertTrue(selection.called);
-		for (int i = 0; i < 10; i++) {
-			// fitnesses of original before selection.
-			assertEquals(expected[9-i]+10, pop.getFitness(i));
-			// subject to mutation to opposite order since we selected, which reversed.
-			assertEquals(expected[i], pop.get(i).id);
-			
-		}
-		assertEquals(16, pop.getFitnessOfMostFit());
-		assertEquals(94, pop.getMostFit().getCost());
-		assertEquals(6, pop.getMostFit().getSolution().id);
-		assertEquals(tracker.getSolution(), pop.getMostFit().getSolution());
-		pop.replace();
-		pop.select();
-		for (int i = 0; i < 10; i++) {
-			assertEquals(expected[9-i], pop.get(i).id);
-			assertEquals(expected[i]+10, pop.getFitness(i));
-		}
-		assertEquals(16, pop.getFitnessOfMostFit());
-		assertEquals(94, pop.getMostFit().getCost());
-		assertEquals(6, pop.getMostFit().getSolution().id);
-		assertEquals(tracker.getSolution(), pop.getMostFit().getSolution());
-		
-		f.changeFitness(1);
-		pop.updateFitness(0);
-		f.changeFitness(10);
-		pop.updateFitness(1);
-		pop.replace();
-		assertEquals(expected[9]+10+1, pop.getFitness(0));
-		assertEquals(expected[8]+10+10, pop.getFitness(1));
-		assertEquals(22, pop.getFitnessOfMostFit());
-		assertEquals(2, pop.getMostFit().getSolution().id);
-		assertEquals(98, pop.getMostFit().getCost());
-		
-		f.changeFitness(12);
-		BasePopulation.Integer<TestObject> pop2 = pop.split();
-		
-		// orginal should be same
-		assertEquals(expected[9]+10+1, pop.getFitness(0));
-		assertEquals(expected[8]+10+10, pop.getFitness(1));
-		assertEquals(22, pop.getFitnessOfMostFit());
-		assertEquals(2, pop.getMostFit().getSolution().id);
-		assertEquals(98, pop.getMostFit().getCost());
-		
-		// trackers should be same
-		assertTrue(pop.getProgressTracker() == pop2.getProgressTracker());
-		
-		pop2.init();
-		assertEquals(10, pop2.size());
-		assertEquals(10, pop2.mutableSize());
-		for (int i = 0; i < 10; i++) {
-			assertEquals(1-i+12+10, pop2.getFitness(i));
-		}
-		
-		assertFalse(pop.evolutionIsPaused());
-		assertFalse(pop2.evolutionIsPaused());
-		tracker.stop();
-		assertTrue(pop.evolutionIsPaused());
-		assertTrue(pop2.evolutionIsPaused());
-		tracker.start();
-		assertFalse(pop.evolutionIsPaused());
-		assertFalse(pop2.evolutionIsPaused());
-		tracker.update(0, new TestObject(), true);
-		assertTrue(pop.evolutionIsPaused());
-		assertTrue(pop2.evolutionIsPaused());	
+		verifyInteger(pop, f, tracker, selection, p -> ((BasePopulation.Integer<TestObject>)p).getFitnessOfMostFit(), 0);
 	}
 	
 	@Test
@@ -677,182 +312,6 @@ public class BasePopulationTests {
 			tracker, 
 			0
 		);
-		pop.init();
-		pop.select();
-		TestObject[] firstSelect = new TestObject[10];
-		for (int i = 0; i < 10; i++) {
-			firstSelect[i] = pop.get(i);
-		}
-		pop.replace();
-		pop.select();
-		TestObject[] secondSelect = new TestObject[10];
-		for (int i = 0; i < 10; i++) {
-			secondSelect[i] = pop.get(i);
-		}
-		for (int i = 0; i < 10; i++) {
-			assertFalse(firstSelect[i] == secondSelect[9-i]);
-			assertEquals(firstSelect[i], secondSelect[9-i]);
-		}
-	}
-
-	
-	private static class TestSelectionOp implements SelectionOperator {
-		
-		boolean called;
-		public TestSelectionOp() {
-			called = false;
-		}
-		@Override
-		public void select(PopulationFitnessVector.Integer fitnesses, int[] selected) {
-			int next = selected.length - 1;
-			for (int i = 0; i < selected.length; i++) {
-				selected[i] = next;
-				next--;
-			}
-			called = true;
-		}
-		@Override
-		public void select(PopulationFitnessVector.Double fitnesses, int[] selected) {
-			int next = selected.length - 1;
-			for (int i = 0; i < selected.length; i++) {
-				selected[i] = next;
-				next--;
-			}
-			called = true;
-		}
-		@Override
-		public TestSelectionOp split() {
-			return new TestSelectionOp();
-		}
-	}
-	
-	private static class TestFitnessDouble implements FitnessFunction.Double<TestObject> {
-		
-		private TestProblemDouble problem;
-		private int adjustment;
-		
-		public TestFitnessDouble() {
-			problem = new TestProblemDouble();
-		}
-		
-		public double fitness(TestObject c) {
-			return c.id + 0.4 + adjustment;
-		}
-		
-		public Problem<TestObject> getProblem() {
-			return problem;
-		}
-		
-		public void changeFitness(int adjustment) {
-			this.adjustment = adjustment;
-		}
-	}
-	
-	private static class TestFitnessDoubleIntCost implements FitnessFunction.Double<TestObject> {
-		
-		private TestProblemInteger problem;
-		private int adjustment;
-		
-		public TestFitnessDoubleIntCost() {
-			problem = new TestProblemInteger();
-		}
-		
-		public double fitness(TestObject c) {
-			return c.id + 10 + adjustment;
-		}
-		
-		public Problem<TestObject> getProblem() {
-			return problem;
-		}
-		
-		public void changeFitness(int adjustment) {
-			this.adjustment = adjustment;
-		}
-	}
-	
-	private static class TestFitnessInteger implements FitnessFunction.Integer<TestObject> {
-		
-		private TestProblemInteger problem;
-		private int adjustment;
-		
-		public TestFitnessInteger() {
-			problem = new TestProblemInteger();
-		}
-		
-		public int fitness(TestObject c) {
-			return c.id + 10 + adjustment;
-		}
-		
-		public Problem<TestObject> getProblem() {
-			return problem;
-		}
-		
-		public void changeFitness(int adjustment) {
-			this.adjustment = adjustment;
-		}
-	}
-	
-	private static class TestProblemInteger implements IntegerCostOptimizationProblem<TestObject> {
-		public int cost(TestObject c) {
-			return 100 - c.id;
-		}
-		public int value(TestObject c) {
-			return cost(c);
-		}
-	}
-	
-	private static class TestProblemDouble implements OptimizationProblem<TestObject> {
-		public double cost(TestObject c) {
-			return 1.0 / (1.0 + c.id);
-		}
-		public double value(TestObject c) {
-			return cost(c);
-		}
-	}
-	
-	private static class TestInitializer implements Initializer<TestObject> {
-		public TestObject createCandidateSolution() {
-			return new TestObject();
-		}
-		public TestInitializer split() {
-			return this;
-		}
-	}
-	
-	private static class TestObject implements Copyable<TestObject> {
-		
-		private static int IDENTIFIER = 0;
-		private static boolean increase = true;
-		public static void reinit() {
-			increase = true;
-			IDENTIFIER = 0;
-		}
-		
-		private int id;
-		
-		public TestObject() {
-			if (increase) IDENTIFIER++;
-			else IDENTIFIER--;
-			id = IDENTIFIER;
-			if (IDENTIFIER == 6) increase = false;
-		}
-		
-		private TestObject(int id) {
-			this.id = id;
-		}
-		
-		public TestObject copy() {
-			return new TestObject(id);
-		}
-		
-		@Override
-		public int hashCode() {
-			return id;
-		}
-		
-		@Override
-		public boolean equals(Object other) {
-			return id == ((TestObject)other).id;
-		}
+		verifySelectCopies(pop);
 	}
 }
